refactor: rename file and some variables
