+++ site/src/layout/PageLayout.tsx
@@ -1,4 +1,4 @@
-import { gridTotalColumns } from "@src/theme";
+import { pageLayoutGridTotalColumns } from "@src/theme";
 import { HTMLAttributes } from "react";
 import styled, { css } from "styled-components";
 
@@ -9,13 +9,13 @@ interface PageLayoutProps extends HTMLAttributes<HTMLDivElement> {
 export const PageLayout = ({ grid, ...restProps }: PageLayoutProps) => <Root $grid={grid} {...restProps} />;
 
 const Root = styled.div<{ $grid?: boolean }>`
-    max-width: ${({ theme }) => theme.defaultMaxWidth};
+    max-width: ${({ theme }) => theme.pageLayoutDefaultMaxWidth};
     margin: 0 auto;
 
     ${({ $grid }) =>
         $grid &&
         css`
             display: grid;
-            grid-template-columns: repeat(${gridTotalColumns}, 1fr);
+            grid-template-columns: repeat(${pageLayoutGridTotalColumns}, 1fr);
         `}
 `;
