+++ site/src/theme.ts
@@ -5,7 +5,7 @@ const createBreakpoint = (value: number) => {
     };
 };
 
-export const gridTotalColumns = 24;
+export const pageLayoutGridTotalColumns = 24;
 
 export const theme = {
     palette: {
@@ -33,7 +33,7 @@ export const theme = {
         },
     },
     fontFamily: "Arial, sans-serif",
-    defaultMaxWidth: "1920px",
+    pageLayoutDefaultMaxWidth: "1920px",
     breakpoints: {
         xs: createBreakpoint(600),
         sm: createBreakpoint(900),
